requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.8.0
aiosqlite>=0.17.0
python-dotenv>=0.19.0
pycoingecko>=3.0.0